Make tests use pytest, added docs
